move data calls to one file
