+++ pages/03_chronic_conditions.py
@@ -1,86 +1,7 @@
 import streamlit as st
-import pandas as pd
 import plost
-import util
 import components
-
-conn = util.connection(database="dev_lipsa")
-
-
-@st.cache_data
-def year_months():
-    query = """
-        select distinct
-           year(claim_end_date)::text || '-' ||
-             lpad(month(claim_end_date)::text, 2, '0')
-             as year_month
-           , sum(paid_amount)
-        from core.medical_claim
-        group by 1
-        having sum(paid_amount) > 10
-        order by 1
-    """
-    data = util.safe_to_pandas(conn, query)
-    return data
-
-
-@st.cache_data
-def pmpm_by_chronic_condition():
-    query = """
-        with conditions as (
-            select distinct
-                year(condition_date)::text || '-' || lpad(month(condition_date)::text, 2, '0') as year_month
-                , claim_id
-                , patient_id
-                , code
-                , condition
-                , condition_family
-            from core.condition
-            inner join chronic_conditions._value_set_tuva_chronic_conditions_hierarchy vs on condition.code = vs.icd_10_cm_code
-            where code_type = 'icd-10-cm'
-        )
-        , medical_spend as (
-            select
-                year(claim_start_date)::text || '-' || lpad(month(claim_start_date)::text, 2, '0') as year_month
-                , claim_id
-                , patient_id
-                , sum(paid_amount) as medical_paid_amount
-            from core.medical_claim
-            group by 1, 2, 3
-        ), merged as (
-            select
-                year_month
-                , condition_family
-                , sum(medical_paid_amount) as medical_paid_amount_sum
-            from conditions
-            join medical_spend using(patient_id, claim_id, year_month)
-            group by 1, 2
-        )
-        select
-            *
-        from merged
-        join pmpm._int_member_month_count using(year_month)
-        order by 2, 1
-    """
-    data = util.safe_to_pandas(conn, query)
-    return data
-
-
-@st.cache_data
-def condition_data():
-    query = """SELECT
-                CONCAT(date_part(year, FIRST_DIAGNOSIS_DATE), '-', lpad(date_part(month, FIRST_DIAGNOSIS_DATE), 2, 0)) AS DIAGNOSIS_YEAR_MONTH,
-                CONDITION,
-                COUNT(*) AS CONDITION_CASES,
-                AVG(LAST_DIAGNOSIS_DATE + 1 - FIRST_DIAGNOSIS_DATE) AS DIAGNOSIS_DURATION
-              FROM CHRONIC_CONDITIONS.TUVA_CHRONIC_CONDITIONS_LONG
-              GROUP BY 1,2
-              ORDER BY 3 DESC;"""
-    data = util.safe_to_pandas(conn, query)
-    data["diagnosis_year"] = pd.to_datetime(
-        data["diagnosis_year_month"]
-    ).dt.year.astype(str)
-    return data
+import data
 
 
 ## --------------------------------- ##
@@ -94,13 +15,13 @@ is `Cardiovascular disease`. The second highest driver is `Metabolic Disease`.
 """
 )
 
-year_month_values = year_months()
+year_month_values = data.year_months()
 year_values = sorted(list(set([x[:4] for x in year_month_values["year_month"]])))
 selected_range = components.year_slider(year_values)
 
 
-chronic_condition_counts = condition_data()
-chronic_condition_data = pmpm_by_chronic_condition()
+chronic_condition_counts = data.condition_data()
+chronic_condition_data = data.pmpm_by_chronic_condition()
 chronic_condition_data = chronic_condition_data.loc[
     chronic_condition_data["year_month"].str[:4].isin(selected_range)
 ]
